+++ react/components/ProductKitDetails.js
@@ -1,32 +1,27 @@
-import PropTypes from 'prop-types'
 import React, { Component } from 'react'
+import PropTypes from 'prop-types'
 import { FormattedMessage } from 'react-intl'
 import { BuyButton, ProductPrice } from 'vtex.store-components'
-import { path } from 'ramda'
 
-import propTypes from '../props/productKitItemProps'
+import { productShape } from '../props'
 
 /**
- * Product Kit Details component.
+ * ProductKitDetails component.
+ *
  * Show the details (price and number of items) of the Kit.
  */
 export default class ProductKitDetails extends Component {
   static propTypes = {
-    /** Array of items which composes the kit */
-    items: PropTypes.arrayOf(propTypes.product).isRequired,
-  }
-
-  static defaultProps = {
-    items: [],
+    /** Array of items */
+    items: PropTypes.arrayOf(
+      productShape,
+    ),
   }
 
   getPrice = item => {
-    return path(['sku', 'seller', 'commertialOffer', 'Price'], item)
+    return item.sku.seller.commertialOffer.Price
   }
 
-  /**
-   * Calculates the Kit Price according to the products that are being displayed.
-   */
   calculateListPrice = items => {
     return items.reduce((price, item) => {
       return price + this.getPrice(item)
@@ -40,7 +35,7 @@ export default class ProductKitDetails extends Component {
   calculateSellingPrice = items => {
     return items.reduce((price, item) => {
       return (
-        price + (this.getPrice(item) * (100 - path(['discount'], item))) / 100
+        price + (this.getPrice(item) * (100 - item.discount)) / 100
       )
     }, 0)
   }
@@ -51,9 +46,9 @@ export default class ProductKitDetails extends Component {
    */
   getSkuItems = items => {
     return items.map(item => ({
-      skuId: String(path(['sku', 'itemId'], item)),
-      quantity: path(['minQuantity'], item),
-      seller: parseInt(path(['sku', 'seller', 'sellerId'], item)),
+      skuId: String(item.sku.itemId),
+      quantity: item.minQuantity,
+      seller: parseInt(item.sku.seller.sellerId),
     }))
   }
 
@@ -61,22 +56,25 @@ export default class ProductKitDetails extends Component {
     const { items } = this.props
 
     return (
-      <div className="vtex-product-kit__details flex flex-column items-center justify-center mw5 mh7">
-        <FormattedMessage
-          id="productKit.numberOfProductsTitle"
-          values={{ numberOfItems: items.length }}
-        />
+      <div className="vtex-product-kit-list__details flex flex-column items-center justify-center tc ph7">
+        <div className="t-body c-muted-1 mv3">
+          <FormattedMessage
+            id="productKitList.takeAll"
+            values={{ numberOfItems: items.length }}
+          />
+        </div>
         <div className="pv4">
           <ProductPrice
             sellingPrice={this.calculateSellingPrice(items)}
             listPrice={this.calculateListPrice(items)}
             showInstallments={false}
+            showLabels={false}
           />
         </div>
         <BuyButton skuItems={this.getSkuItems(items)}>
-          <FormattedMessage id="productKit.buyTogether" />
+          <FormattedMessage id="productKitList.buyKit" />
         </BuyButton>
-      </div>
+      </div >
     )
   }
 }