+++ react/helpers/index.js
@@ -1,19 +1,25 @@
 import { propEq, findIndex } from 'ramda'
 
 /**
- * Helper functions to extract the Kit Items of a Product Kit.
+ * Helper functions to extract the Kit Items of a Product Kit and sets the first
+ * item as the main item according to the mainProductId passed as an argument.
  */
-export const extractItemsKit = productsKit => {
-  let itemsKit = []
+export const extractItemsKit = (productsKit, mainProduct) => {
+  const itemsKit = []
+  let indexOfMainProduct = -1
 
   if (productsKit) {
-    productsKit.forEach(productKit => {
+    productsKit.forEach((productKit, index) => {
       const {
         benefitProduct,
         benefitSKUIds,
         discount,
         minQuantity,
       } = productKit
+
+      if (benefitProduct.productId === mainProduct.productId && indexOfMainProduct === -1) {
+        indexOfMainProduct = index
+      }
 
       benefitSKUIds.forEach(skuId => {
         const indexOfItem = findIndex(propEq('itemId', skuId))(
@@ -34,6 +40,12 @@ export const extractItemsKit = productsKit => {
         }
       })
     })
+  }
+
+  if (indexOfMainProduct !== -1) {
+    const mainItem = itemsKit[indexOfMainProduct]
+    itemsKit[indexOfMainProduct] = itemsKit[0]
+    itemsKit[0] = mainItem
   }
 
   return itemsKit
@@ -70,8 +82,7 @@ export const extractKitItem = productKit => {
       )
     }
 
-    kitItem.sku.referenceId = (kitItem.sku.referenceId &&
-      kitItem.sku.referenceId[0]) || {
+    kitItem.sku.referenceId = (kitItem.sku.referenceId && kitItem.sku.referenceId[0]) || {
       Value: '',
     }
   }