+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- New design.
+- New feature of Item Kit Swap.
+- New feature of Item Kit Removal.
+- Display the base product as the first in every product kits.
+- Design tokens.
 
 ## [1.0.0] - 2018-11-05